% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regress.R
\name{predict.regress}
\alias{predict.regress}
\title{predict.regress function}
\usage{
\method{predict}{regress}(object, newdata, X = NULL, Y = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{regress}}

\item{newdata}{optional new data for prediction in the same format
as \code{X} used in the original \code{regress} call}

\item{X}{original predictors from the \code{regress} call. Needed
for some prediction algorithms when using \code{newdata}}

\item{Y}{original dependent variable from \code{regress}
call. Needed for some prediction algorithms when using
\code{newdata}}

\item{...}{ignored}
}
\value{
<return>
}
\description{
predict.regress function
}
\author{
Sam Lendle
}

