\name{regress}
\alias{regress}
\title{<title>}
\usage{regress(Y, X, family=binomial(), method="glm", formula=Y ~ ., ...)}
\description{<description>}
\details{<details>}
\value{<return>}
\author{Sam Lendle}
\arguments{\item{Y}{}
\item{X}{}
\item{family}{}
\item{method}{}
\item{formula}{}
\item{...}{}}
