\name{setColnames}
\alias{.setColnames}
\alias{setColnames}
\title{<title>}
\usage{
  .setColnames(x.colnames, x.ncols, firstChar)
}
\arguments{
  \item{x.colnames}{<description>}

  \item{x.ncols}{<description>}

  \item{firstChar}{<description>}
}
\value{
  <return>
}
\description{
  <description>
}
\details{
  <details>
}
\author{
  Susan Gruber
}

