\name{predict.regress}
\alias{predict.regress}
\title{<title>}
\usage{predict.regress(object, newdata, X, Y, ...)}
\description{<description>}
\details{<details>}
\value{<return>}
\author{Sam Lendle}
\arguments{\item{object}{an object of class \code{regress}}
\item{newdata}{optional new data for prediction in the same format
as \code{X} used in the original \code{regress} call}
\item{X}{original predictors from the \code{regress} call. Needed
for some prediction algorithms when using \code{newdata}}
\item{Y}{original dependent variable from \code{regress}
call. Needed for some prediction algorithms when using
\code{newdata}}
\item{...}{ignored}}
