\name{print.cte}
\alias{print.cte}
\title{Prints object of class \code{cte}}
\usage{
  \method{print}{cte} (x, ...)
}
\arguments{
  \item{x}{an ojbect of class \code{cte}}

  \item{...}{ignored}
}
\value{
  Nothing
}
\description{
  Prints results from tmle.cte.
}
\author{
  Sam Lendle \email{lendle@stat.berkeley.edu}
}

