% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmlecte.R
\name{print.cte}
\alias{print.cte}
\title{\code{print.cte} function}
\usage{
\method{print}{cte}(x, ...)
}
\arguments{
\item{x}{an ojbect of class \code{cte}}

\item{...}{ignored}
}
\value{
Nothing
}
\description{
Prints results from tmle.cte.
}
\author{
Sam Lendle
}

