% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setColnames}
\alias{.setColnames}
\alias{setColnames}
\title{.setColnames function}
\usage{
.setColnames(x.colnames, x.ncols, firstChar)
}
\arguments{
\item{x.colnames}{current column names}

\item{x.ncols}{current number of columns}

\item{firstChar}{prefix for internally assigned name}
}
\value{
column names
}
\description{
.setColnames function
}
\details{
assign names to every unnamed column of x
}
\author{
Susan Gruber
}

