\name{setColnames}
\title{<title>}
\usage{.setColnames(x.colnames, x.ncols, firstChar)}
\description{<description>}
\details{<details>}
\value{<return>}
\author{Susan Gruber}
\alias{.setColnames}
\arguments{\item{x.colnames}{<description>}
\item{x.ncols}{<description>}
\item{firstChar}{<description>}}
