% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tmle.cte}
\alias{tmle.cte}
\title{tmle.cte function}
\usage{
tmle.cte(A, B, Y, a = 0, Delta = NULL, Q.method = "glm",
  Q.formula = NULL, Q.SL.library = NULL, Q.A1 = NULL, Q.A0 = NULL,
  g.method = "glm", g.formula = NULL, g.SL.library = c("SL.glm",
  "SL.step", "SL.knn"), g.A1 = NULL, gDelta.method = "glm",
  gDelta.formula = NULL, gDelta.SL.library = c("SL.glm", "SL.step",
  "SL.knn"), gDelta.1 = NULL, family = gaussian(), tol = 1e-10,
  maxiter = 100, target = TRUE, verbose = FALSE, Qbound = c(1e-10, 1 -
  1e-10), gbound = c(1e-10, 1 - 1e-10), ...)
}
\arguments{
\item{A}{binary treatment indicator where 1 is treated and 0 is
not treated}

\item{B}{a matrix or dataframe of covariates.}

\item{Y}{outcome variable. Currently should be in [0, 1], but can
be continuous.}

\item{a}{the treatment for which you want to estimate the
conditional effect, 1 or 0. 1 corresponds to the ATT, 0
corresponds to the ATN or NDE.}

\item{Delta}{binary indicator of missing outcome.  Should be 0 where \code{Y} is \code{NA} or should be treated as \code{NA}, and 1 otherwise.}

\item{Q.method}{method to estimate \code{Q(A,B)=E(Y|A,B)}. Should be \code{"glm"} for glm,
\code{"SL"} for SuperLearner, or \code{"user"} for user specified estimates of
\code{g(A=1|B)}. If SuperLearner is specified but not installed, glm is automatically used.}

\item{Q.formula}{the formula for glm if glm is used to estimate
\code{Q(A,B)}. The response variable should be calle d \code{Y}
and the predictors should include \code{A} and variables in
\code{B}.  If missing, defaults to \code{Y~.} (main terms including
\code{A} all variables in \code{B}).}

\item{Q.SL.library}{SuperLearner library for estimating
\code{Q(A,B)} if SuperLearner is used.}

\item{Q.A1}{User specified estimates of \code{Q(1|B)}.  Should be the same length as
\code{Y} if \code{Q.method} is \code{"user"}.}

\item{Q.A0}{User specified estimates of \code{Q(0|B)}.  Should be the same length as
\code{Y} if \code{Q.method} is \code{"user"}.}

\item{g.method}{method to estimate
\code{g(A=a|B)=P(A=a|B)}. Should be \code{"glm"} for glm,
\code{"SL"} for SuperLearner, or \code{"user"} for user specified estimates of
\code{g(A=1|B)}. If SuperLearner is specified but not installed, glm is automatically used.}

\item{g.formula}{the formula for glm if glm is used to estimate
\code{g(A=a|B)}. The response variable should be calle d \code{A}
and the predictors should include variables in \code{B}.  If
missing, defaults to $A~.$ (main terms including all variables in
\code{B}).}

\item{g.SL.library}{SuperLearner library for estimating
\code{g(A=a|B)} if SuperLearner is used.}

\item{g.A1}{User specified estimates of \code{g(A=1|B)}.  Should be the same length as
\code{Y} if \code{g.method} is \code{"user"}. Note: this is the estimate for \code{g(A=1|B)}
even if \code{a=0} (e.g. for the NDE or for the ATN).}

\item{gDelta.method}{method to estimate
\code{P(Delta=1|A, B)}. Should be \code{"glm"} for glm,
\code{"SL"} for SuperLearner, or \code{"user"} for user specified estimates of
\code{g(A=1|B)}. If SuperLearner is specified but not installed, glm is automatically used.}

\item{gDelta.formula}{the formula for glm if glm is used to
estimate \code{P(Delta=1|A, B)}. The response variable should be
called \code{Delta} and the predictors should include \code{A} and
variables in \code{B}.  If missing, defaults to $Delta~.$ (main
terms including \code{A} and all variables in \code{B}).}

\item{gDelta.SL.library}{SuperLearner library for estimating
\code{P(Delta=1|A,B)} if SuperLearner is used.}

\item{gDelta.1}{User specified estimates of \code{P(Delta=1|A,B)}.  Should be the same length as
\code{Y} if \code{gDelta.method} is \code{"user"}.}

\item{family}{\code{gaussian} for a continuous outcome \code{Y},
\code{binomial} for a biniomial outcome}

\item{tol}{convergence criterion, set to something small}

\item{maxiter}{maximum number of targetting steps. 100 is good. If
it takes more than that it's probably not going to converge.}

\item{target}{TRUE to preform bias reduction (targeting)
step. FALSE to return simple G computation based substitution
estimate}

\item{verbose}{TRUE for details for each iteration}

\item{Qbound}{a numeric vector containing minimum and maximum
bounds for Q. Should be between 0 and 1 for now.  Set to something
close to 0 and 1.}

\item{gbound}{a numeric vector containing minimum and maximum
bounds for g. Should be between 0 and 1. Set to something close to
0 and 1.}

\item{...}{aditional parameters to be passed to (both)
perLearner calls}
}
\value{
An object of class \code{cte}
}
\description{
Estimates the ``conditional treatment effect'' E(E(Y|A=1,B)-E(Y|A=0,B)|A=a) using Targetted Maximum Likelihood.
}
\author{
Sam Lendle \email{lendle@stat.berkeley.edu}
}

